refactor(data): Change coin quantity from Double to Int
This commit refactors the data model and corresponding UI components to treat the quantity of gold coins as an integer (`Int`) instead of a floating-point number (`Double`), ensuring that investments are tracked in whole units.

**Key Changes:**

*   **Data Model (`GoldCoin.kt`)**:
    *   The `quantity` property in the `GoldCoin` data class is changed from `Double` to `Int`. This enforces that coins can only be held in whole numbers.

*   **ViewModel (`GoldViewModel.kt`)**:
    *   The `insert` function's signature is updated to accept `qty` as an `Int`.

*   **UI (`MainScreen.kt`, `DetailScreen.kt`)**:
    *   The "Add Coin" dialog now accepts only numeric digits for the quantity input, parsing the value to an `Int`.
    *   Display labels have been updated from "units" to the more specific "coins" for clarity (e.g., "10 coins").
    *   Input field labels in the dialog are now more descriptive (e.g., "Number of Coins," "Price per Coin").

+++ app/src/main/java/com/example/portfolioapp/data/GoldCoin.kt
@@ -7,11 +7,11 @@ import androidx.room.PrimaryKey
 data class GoldCoin(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val name: String,
-    val originalPrice: Double, // Price PER COIN
-    val currentPrice: Double,  // Price PER COIN
-    val quantity: Double       // Number of coins
+    val originalPrice: Double,
+    val currentPrice: Double,
+    val quantity: Int
 ) {
-    // Helper properties (computed on the fly)
+    // Math automatically works (Int * Double = Double)
     val totalCurrentValue: Double
         get() = currentPrice * quantity
